SOPs/localllm-notebooklm/requirements.txt
python-dotenv
google-genai
openai
